Rename package name for TMailPathConverterTest to Apache James one
Lazy way to avoid changing an access modifier on Apache James test suite.
